NOMDL Workshop: Improved detection of filled description and instructions
The function strip_tags() used to be here as a way around a bug in
TinyMCE producing empty <br /> instead of empty string. This lead to the
unwanted behaviour that Workshop description of instructions consisting
of <img> only were considered as empty. The TinyMCE seems to produce
correct string now so this is not needed any more.