question bank MDL-24312 problem when deleting a course with orphaned categories.
When there was a question category whose parent had somehow been deleted from the question bank, while the child category was still there, then that gave an error because a DB query was still referring to the question_categories.courese coloumn, that was changed to contextid in moodle 1.9.

Fix thanks to Mahmoud Kassaei.