MDL-25886 allow user_picture::fields() to be used multiple times within a single query
When user table is joined twice or more within a single query, it is
important to be able to provide aliases to all returned files. This
patch adds optional parameter that, if provided, is used as a prefix
for columns aliases. Does not apply to the 'id' column, use $idalias for
that.