MDL-25523 Workshop: not-grouped users are excluded from allocation in visible group mode
There was a bug as these users were ignored only as squares but not as
circles. This patch makes sure that no new allocation are added to
not-grouped users in visible group mode. However, the options 'Remove current
allocations' and 'Add self-assessments' apply to them intentionally.

Also, the user is warned if there are some not-grouped users found in a
workshop in visible groups mode or separate groups mode.