Fixed import groups page layout and navigation
While working on MDL-26252, I realized that group/import.php page had
issues with layout. Side blocks were missing and the breadcrumb navigation
broken (leading to a non-existing script and having the link to the
current course duplicated).

This patch sets the standard layout for the page so that block are
displayed, fixes the page position in the navigation tree and the
breadcrumb navigation.