MDL-26312 Workshop: data integrity check to detect eventual course_modules corruption
Because of a bug in Workshop upgrade code, multiple workshop course_modules can
potentially point to a single workshop instance. The chance is pretty low as in most cases,
the upgrade failed. But under certain circumstances, workshop could be upgraded with
this data integrity issue. We want to detect it now and let the admin know.