MDL-27065 workshop grades categories can be defined in mod_form
Because the Workshop module creates two grade items in the gradebook, we
can't use standard modedit features to set the corresponding category in
the gradebook (it supports single grade item only).

The patch also cleans the mod_form.php so that it does not use the
global $COURSE (just because I do not like this global variable).

And for the record - QuickForms must die.