MDL-33330 introducing new file_reference_exception
This replaces the previously defined external_file_exception that was
not used anywhere in the code. The new exception will be thrown in
situations where there is some unexpected problem detected with the
records in the {files_reference} table (e.g. unable to find a record
referenced by some file from {file}, invalid reference value etc.)