MDL-35238 Introduce a first version of the mdeploy.php script
This is supposed to serve as a standalone script that accepts parameters
via CLI or HTTP. As no Moodle library can be reliably included, we have
to implement our own input handling, output handling and processing the
actual update/install task.

PHPUnit is used to test the functionality. Although (again), the testing
must not be included as a part of standard Moodle PHPUnit environment
but explicitly.