MDL-42039 Restore: Prevent adding too many form fields at once
This change relates to the Moodle forms used to select activities
in the restore process.

If you add too many form fields at once, it takes a very long time
during which we can't report progress. But adding fields in smaller
groups takes longer overall, so it's a tough balance. This change
splits the adds into groups of 1,000, reporting progress between each.