MDL-41761 statistics should be calculable for last or highest attempts
as well as for first or all attempts.

This commit drops a database table and then recreates it. This is in
order to be able to have some new columns in the table be NOT NULL with
no default value, which would not be possible if I added fields to an
existing table.

The quiz_statistics table is used for caching calculated values only
and thus we can safely drop it and recreate it.