MDL-43104 JavaScript: Combine mcore JS into simplemoodle
I was wrong about why this didn't work before - this was caused by the
missing semi-colon.

Joining these scripts into a single combo-load has the benefit of further
reducing the number of HTTP transactions to load a page, and a minor
reduction on transferred size due to a reduced header count.